Add a simple example that shows how to use the generated types
The example-macro sub-project imports an example.json that provides a
good use case to start with. The main.rs just had `import_types!` invocation
which will generate the rust type. This PR adds on to that by showing how the
generated types can be used.

This more of a documentation PR than an actual bug. This fixes #14