Support telling serde where it is, i.e. when re-exported as a transitive dependency.
Supported by the `serde()` annotation in derive macros, which otherwise
will assume that serde is a direct crate dependency. When a macro emits
a type that derives Serialize/Deserialize, the user of that macro would
normally have to directly depend on serde themselves for the resulting
expansion of those derive macros. This can be cumbersome when consuming
serde transitively via some other crate (i.e. progenitor).

https://github.com/serde-rs/serde/blob/5a8dcac2ed1407fab3f7fd23f2d56af42dcd448f/serde_derive/src/internals/attr.rs#L556-L561