Better errors for schema merging
I've been debugging errors originating in the schema merge logic
lately, and I thought I could improve the errors somewhat.
This is a bit of a low-effort attempt. Not all the errors are
tremendously useful, and the error messages are mostly guesswork.
But they ought to be somewhat better than what came before.