Decanoncalize std types in the expected and actual ASTs
Based on a PR comment from @ahl.  Remove some ugly special-case
handling, and unconditionally decanonicalize all `std`-prefixed types in
both the generated code as well as the original types' ASTs.  With this
in place, we can expect the two ASTs to match exactly, which some of the
tests assert.