Fix (and test) a missing fully-qualified `String` use
This was in the codegen for struct builders, which `schemas.rs` do not enable by default so it didn't fail with my cursed schema.  This change enables generating struct builders for *all* of the test cases in `schemas.rs`, and verifies that the generated code properly compiles now that the remaining `String` is canonicalized.

This also changes all of the generated tests but that will be in a separate commit for easier reviewing.